Fixed systray
The systray was incorrectly displayed when the bar was drawn:

- The systray width was not added to the status text
- The title width was taking into account the systray width that was not
correctly displayed because of previous condition

The result was that both the title and the status bars were too short, leading
to visual glitches when the screen size changed: neither the title bar nor the
status bar was covering the zone corresponding to the systray width on
the left of the status bar.

This patch fixes the systray patch by adding the systray width to the
status width (`tw`).